byte<->sbyte casts at 6 heading/category/levelRestrict sites: House/Equipment (byte->sbyte), RecallInstantEffect/SM_IN_GAME_SHOP_LIST x2/TeleportService (sbyte->byte) (Java byte=sbyte) (968->962)

+++ dotnetConversion/src/Aion.GameServer/Model/GameObjects/Player/Equipment.cs
@@ -57,7 +57,7 @@ public partial class Equipment : IPersistable
             return null;
         }
 
-        sbyte levelRestrict = itemTemplate.GetMaxLevelRestrict(owner.GetPlayerClass());
+        sbyte levelRestrict = (sbyte)itemTemplate.GetMaxLevelRestrict(owner.GetPlayerClass());
         if (levelRestrict != 0 && owner.GetLevel() > levelRestrict)
         {
             Aion.GameServer.Utils.PacketSendUtility.SendPacket(owner, Aion.GameServer.Network.Aion.ServerPackets.SmSystemMessage.STR_CANNOT_USE_ITEM_TOO_HIGH_LEVEL(levelRestrict, itemTemplate.GetL10n()));
